+++ i18n/locales/en/common.json
@@ -77,7 +77,6 @@
   "Create": "Create",
   "Create Lisk ID": "Create Lisk ID",
   "Create a Lisk ID to gain access to all services.": "Create a Lisk ID to gain access to all services.",
-  "Create a New Lisk Account": "Create a New Lisk Account",
   "Create your Lisk ID": "Create your Lisk ID",
   "Create your second passphrase": "Create your second passphrase",
   "Currency": "Currency",
@@ -139,7 +138,6 @@
   "Height": "Height",
   "Help": "Help",
   "Help Center": "Help Center",
-  "Here you can see all the initialized Lisk Accounts in your Hardware Wallet. You can also create a new one.": "Here you can see all the initialized Lisk Accounts in your Hardware Wallet. You can also create a new one.",
   "Here you see your address, balance, and avatar.": "Here you see your address, balance, and avatar.",
   "Hide passphrase": "Hide passphrase",
   "How do you like the hub?": "How do you like the hub?",
@@ -152,7 +150,6 @@
   "In": "In",
   "Include your operating system and screen resolution in your report": "Include your operating system and screen resolution in your report",
   "Incoming": "Incoming",
-  "Index": "Index",
   "Info": "Info",
   "Initialize Lisk ID": "Initialize Lisk ID",
   "Input": "Input",
@@ -186,7 +183,6 @@
   "Lisk ID": "Lisk ID",
   "Lisk Website": "Lisk Website",
   "Lisk.Chat": "Lisk.Chat",
-  "Loading information from your harware wallet...": "Loading information from your harware wallet...",
   "Local": "Local",
   "Log in": "Log in",
   "Log in with Ledger": "Log in with Ledger",
@@ -322,7 +318,6 @@
   "Submit": "Submit",
   "Success": "Success",
   "Success!": "Success!",
-  "Switch": "Switch",
   "Switch networks (Main-/Testnet, Custom)": "Switch networks (Main-/Testnet, Custom)",
   "Take a tour to see how it works. It should not take longer than 5 minutes": "Take a tour to see how it works. It should not take longer than 5 minutes",
   "Tell us what you think": "Tell us what you think",
@@ -351,7 +346,6 @@
   "Transaction": "Transaction",
   "Transaction ID": "Transaction ID",
   "Transaction is being processed and will be confirmed. It may take up to 15 minutes to be secured in the blockchain.": "Transaction is being processed and will be confirmed. It may take up to 15 minutes to be secured in the blockchain.",
-  "Transactions": "Transactions",
   "Transactions can’t be reversed": "Transactions can’t be reversed",
   "Transfer": "Transfer",
   "Try again": "Try again",
