+++ src/components/ledger/unlockWallet.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { translate } from 'react-i18next';
+
+const UnlockWallet = ({ history, handleOnClick }) => (<div>
+  UnlockWallet
+  <button onClick={() => { handleOnClick(); }} >I have unlocked account</button>
+</div>);
+
+export default translate()(UnlockWallet);
